Add dev console to play around
We are stubbing API response for our test suite, but sometime we
need to interact with the actual API request and it's responses.
This commit adds an easier way to play with the Discount Network
API in your console, to start playing simply run `bin/console`.

To make it more easier, it also provides you a `.sample.pryrc`
and all you need do is copy it as `.pryrc` and set your keys.